Refactor typecard.
Variable printing was put into a function for better readability.
CHAR_BIT is used instead of a literal 8 for better portability. The
system bit-width is now detected by checking the size of a pointer. The
proper preprocesor macro is now used for testing for __int128_t. Endian
testing macros were moved out of main, support for PDP endianness has
been added, and the proper GCC endian preprocessor macros are used.